+++ packages/ui/src/components/NumberInput/__stories__/Template.stories.tsx
@@ -1,8 +1,4 @@
 import type { StoryFn } from '@storybook/react-vite'
 import { NumberInput } from '..'
 
-export const Template: StoryFn<typeof NumberInput> = props => (
-  <>
-    <NumberInput {...props} />
-  </>
-)
+export const Template: StoryFn<typeof NumberInput> = props => <NumberInput {...props} />
